Expose an API for setting the active search result (#641)
