+++ src/apis/index.js
@@ -135,6 +135,7 @@ import setNoteTransformFunction from './setNoteTransformFunction';
 import selectThumbnailPages from './selectThumbnailPages';
 import unselectThumbnailPages from './unselectThumbnailPages';
 import setSearchResults from './setSearchResults';
+import setActiveResult from './setActiveResult';
 
 export default store => {
   window.readerControl = {
@@ -199,6 +200,7 @@ export default store => {
     setZoomLevel,
     setZoomList: setZoomList(store),
     setSearchResults: setSearchResults(store),
+    setActiveResult: setActiveResult(store),
     showErrorMessage: showErrorMessage(store),
     textPopup: textPopup(store),
     toggleElement: toggleElement(store),
