+++ src/redux/actions/exposedActions.js
@@ -238,3 +238,11 @@ export const setSearchResults = searchResults => ({
   type: 'SET_SEARCH_RESULTS',
   payload: searchResults,
 });
+export const setActiveResult = (activeResult, index) => ({
+  type: 'SET_ACTIVE_RESULT',
+  payload: { activeResult },
+});
+export const setActiveResultIndex = index => ({
+  type: 'SET_ACTIVE_RESULT_INDEX',
+  payload: { index },
+});
