+++ src/redux/actions/internalActions.js
@@ -323,14 +323,6 @@ export const setSearchValue = value => ({
   type: 'SET_SEARCH_VALUE',
   payload: { value },
 });
-export const setActiveResult = activeResult => ({
-  type: 'SET_ACTIVE_RESULT',
-  payload: { activeResult },
-});
-export const setActiveResultIndex = index => ({
-  type: 'SET_ACTIVE_RESULT_INDEX',
-  payload: { index },
-});
 export const addResult = result => ({
   type: 'ADD_RESULT',
   payload: { result },
